Allow hoisted route mappings
When we request for a url and it includes more content after a slash in
the format `/{key}/{rest:.*}`, then we assume it was a service hoisted
on this key so we redirect to the mapped url and append the rest to it's
url to forward the request to the service with the rest of the url as
requested. This obviously is only good for GET requests.